Secured and objectified game data
Enclosed the entirety of the game withing a "Game" object, making appropriate
use of private variables in order to prevent intrusion into the game.  Note
that a user can still override the custom "requestAnimFrame" function to
change the speed of gameplay.

Bolstered some of the objects and variable names that are part of the game.
This included renaming a few variables, merging one-or-two-line functions into
their calling functions, and making the flow of the game code more visually
explicit.

BROKEN: Brick colours are not properly displaying now for some reason